Moved storybook stories into each components folder
