+++ web/.storybook/main.js
@@ -1,5 +1,9 @@
 module.exports = {
-  stories: ['../stories/**/*.stories.mdx', '../stories/**/*.stories.@(js|jsx|ts|tsx)'],
+  stories: [
+    '../stories/**/*.stories.mdx',
+    '../stories/**/*.stories.@(js|jsx|ts|tsx)',
+    '../components/*/*.stories.js'
+  ],
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
