+++ web/components/Category/Category.stories.js
@@ -1,4 +1,4 @@
-import { Category } from '../../components/Category';
+import { Category } from '.';
 
 export default {
   title: 'Components/Category',
