+++ web/components/Counter/Counter.stories.js
@@ -1,4 +1,4 @@
-import { Counter } from '../../components/Counter';
+import { Counter } from '.';
 
 export default {
   title: 'Components/Counter',
