+++ web/components/Layout/Layout.stories.js
@@ -1,4 +1,4 @@
-import { Layout } from '../../components/Layout';
+import { Layout } from '.';
 
 export default {
   title: 'Components/Layout',
